rag_cpg_pipeline/requirements.txt
pymupdf
google-auth
jsonlines
pytesseract
Pillow
spacy
sentence-transformers
torch
transformers
einops
git+https://github.com/nomic-ai/megablocks.git
